refactor: extract directive helpers
